Preview renaming; ask  permission to proceed

+++ exifRename.mjs
@@ -14,6 +14,7 @@ import { execSync } from 'child_process';
 import { exit } from 'process';
 import { extname } from 'path';
 import { renameSync, existsSync } from 'fs';
+import readline from 'readline';
 
 const exifCommand = `exiftool -DateTimeOriginal -ee -j -q -d "%Y-%m-%d %H∶%M∶%S" .`;
 const getFileInfoCommand = 'GetFileInfo -d';
@@ -44,6 +45,23 @@ if(total === 0){
 
 let renamed = 0;
 let skipped = 0;
+// filter out unsupported extensions
+list = list.filter(item => {
+  const ext = extname(item.SourceFile).toLowerCase();
+  if(supportedExtensions.includes(ext)){
+    return true;
+  } else {
+    skipped += 1;
+    return false;
+  }
+});
+
+if(list.length === 0){
+  console.error('No supported files found.');
+  process.exit(1);
+}
+
+// first pass: compute frequencies and new file names
 list.forEach(item => {
     // check extension
     const extension = extname(item.SourceFile).toLowerCase();
@@ -74,28 +92,64 @@ list.forEach(item => {
         newFileName = `./${item.DateTimeOriginal} (${frequencies[newFileName]})${extension}`;
     } 
     item.NewFileName = newFileName;
+});
+const firstPassEnd = Date.now();
 
-    // rename files if there's a new name and if it's different
-    if (!item.NewFileName || item.SourceFile === item.NewFileName) {
-      // console.error(`Skipping: ${item.SourceFile}`);
-      skipped += 1;
-      return;
-    }
-    // check if a file with desired new name already exists and bail out, otherwise renameSync would overwrite it
-    if(existsSync(item.NewFileName)){
-        console.error(`Error: A file ${item.NewFileName} already exists!`);
+// filter out files where the new name is the same as old name
+list = list.filter(item => {
+  if(item.SourceFile !== item.NewFileName){
+    return true;
+  } else {
+    skipped += 1;
+    return false;
+  }
+});
+if(list.length === 0) {
+  console.error('All files are properly named. Aborting.');
+  process.exit(1);
+}
+
+// second pass: display renaming preview and ask for confirmation
+console.log('Attempting rename:');
+console.table(list, ['SourceFile', 'NewFileName']);
+const rl = readline.createInterface({
+  input: process.stdin,
+  output: process.stdout
+});
+rl.question('\nProceed? [y/N] ', (answer = 'n') => {
+  if(answer.toLowerCase() === 'y'){
+    // third pass: do rename
+    const renameStartTime = Date.now();
+    list.forEach(item => {
+      // rename files if there's a new name and if it's different
+      if (!item.NewFileName || item.SourceFile === item.NewFileName) {
+        // console.error(`Skipping: ${item.SourceFile}`);
         skipped += 1;
         return;
-    }
-    // finally do rename
-    try {
-        renameSync(item.SourceFile, item.NewFileName);
-        console.log(`Renaming: ${item.SourceFile} → ${item.NewFileName}`);
-        renamed += 1;
-    } catch (error) {
-        skipped += 1;
-        console.error(error);
-    }
-});
-const duration = ((Date.now() - startTime) / 1000).toFixed(1);
-console.log(`${renamed ? '——\n': ''}Renamed: ${renamed}, Skipped: ${skipped}, Total: ${total}.\nTime: ${duration} sec.`);
+      }
+      // check if a file with desired new name already exists and bail out, otherwise renameSync would overwrite it
+      if(existsSync(item.NewFileName)){
+          console.error(`Error: A file ${item.NewFileName} already exists!`);
+          skipped += 1;
+          return;
+      }
+      // finally do rename
+      try {
+          renameSync(item.SourceFile, item.NewFileName);
+          console.log(`Renaming: ${item.SourceFile} → ${item.NewFileName}`);
+          renamed += 1;
+      } catch (error) {
+          skipped += 1;
+          console.error(error);
+      }
+    });
+    const firstPassDuration = (firstPassEnd - startTime) / 1000;
+    const renamePassDuration = (Date.now() - renameStartTime) / 1000;
+    const totalDuration = (firstPassDuration + renamePassDuration).toFixed(1);
+    console.log(`${renamed ? '——\n': ''}Renamed: ${renamed}, Skipped: ${skipped}, Total: ${total}.\nTime: ${totalDuration} sec.`);
+  } else {
+    console.log('Aborted.');
+    process.exit(0);
+  }
+  rl.close();
+})
